more build tag protections
