+++ client/firewall/uspfilter/forwarder/endpoint.go
@@ -1,3 +1,5 @@
+//go:build !js
+
 package forwarder
 
 import (
