+++ client/firewall/uspfilter/forwarder/tcp.go
@@ -1,3 +1,5 @@
+//go:build !js
+
 package forwarder
 
 import (
