+++ client/firewall/uspfilter/forwarder/udp.go
@@ -1,3 +1,5 @@
+//go:build !js
+
 package forwarder
 
 import (
