+++ client/internal/dns/tcpstack.go
@@ -1,3 +1,5 @@
+//go:build !js
+
 package dns
 
 import (
